ERA-13041: Rename the fields folder to formElements for consistency and improved semantics.
