+++ src/ReportManager/DetailsSection/SchemaForm/formElements/Boolean/index.js
