+++ src/ReportManager/DetailsSection/SchemaForm/formElements/Boolean/styles.module.scss
