+++ src/ReportManager/DetailsSection/SchemaForm/formElements/ChoiceList/List.js
