+++ src/ReportManager/DetailsSection/SchemaForm/formElements/ChoiceList/index.js
