+++ src/ReportManager/DetailsSection/SchemaForm/formElements/ChoiceList/styles.module.scss
