+++ src/ReportManager/DetailsSection/SchemaForm/formElements/Collection/SortableList/Item/FormModal/index.js
