+++ src/ReportManager/DetailsSection/SchemaForm/formElements/Collection/SortableList/Item/FormModal/styles.module.scss
