+++ src/ReportManager/DetailsSection/SchemaForm/formElements/Collection/SortableList/Item/FormPreview/index.js
