+++ src/ReportManager/DetailsSection/SchemaForm/formElements/Collection/SortableList/Item/FormPreview/styles.module.scss
