+++ src/ReportManager/DetailsSection/SchemaForm/formElements/Collection/SortableList/Item/index.js
