+++ src/ReportManager/DetailsSection/SchemaForm/formElements/Collection/SortableList/Item/styles.module.scss
