+++ src/ReportManager/DetailsSection/SchemaForm/formElements/Collection/SortableList/Item/utils/index.js
