+++ src/ReportManager/DetailsSection/SchemaForm/formElements/Collection/SortableList/SortableItem/index.js
