+++ src/ReportManager/DetailsSection/SchemaForm/formElements/Collection/SortableList/index.js
