+++ src/ReportManager/DetailsSection/SchemaForm/formElements/Collection/index.js
