+++ src/ReportManager/DetailsSection/SchemaForm/formElements/Collection/styles.module.scss
