+++ src/ReportManager/DetailsSection/SchemaForm/formElements/DateTime/styles.module.scss
