+++ src/ReportManager/DetailsSection/SchemaForm/formElements/Header/index.js
