+++ src/ReportManager/DetailsSection/SchemaForm/formElements/Header/styles.module.scss
