+++ src/ReportManager/DetailsSection/SchemaForm/formElements/Location/styles.module.scss
