+++ src/ReportManager/DetailsSection/SchemaForm/formElements/Numeric/index.js
