+++ src/ReportManager/DetailsSection/SchemaForm/formElements/Numeric/styles.module.scss
