+++ src/ReportManager/DetailsSection/SchemaForm/formElements/Section/index.js
