+++ src/ReportManager/DetailsSection/SchemaForm/formElements/Section/styles.module.scss
