+++ src/ReportManager/DetailsSection/SchemaForm/formElements/Text/styles.module.scss
